requirements.txt
django
djangorestframework
celery
redis
psycopg2-binary
django-celery-beat
python-decouple
